lightning: move test case to schema_import_test.go
